requirements/read_flow.txt
keyboard~=0.13.5
requests~=2.25.0
google-api-python-client~=1.12.8
google-auth-httplib2~=0.0.4
google-auth-oauthlib~=0.4.2
googletrans~=3.0.0
pycairo~=1.20.0
PyGObject~=3.38.0
rich~=9.2.0
click~=7.1.2
cryptography==3.2.1